stumpy: update asound.state for PVT build.
The new pin config in the Cirrus driver removes controls for
unconnected paths.  Mic is now attached to line in.  Mic in is not
attached, DMIC is not attached, neither is SPDIF out. remove these
from asound.state as their controls will no longer be present.

BUG=chrome-os-partner:6135
TEST=sound_infrastructure, plus manually boot stumpy and check with
arecord and aplay. Needs changes: 11189, 11194

Signed-off-by: Dylan Reid <dgreid@chromium.org>
Change-Id: Idfda648d3a6a87f8fffb5c9315e0dddf6de99925