CRAS: Add alsa_io device.
This is the heart of the server.  It communicates with an alsa device,
managing wakeup timing and forwarding audio samples to/from clients.

A thread is created for the device.  For playback this thread will
wake up when the buffer starts to run low, ask the client(s) for more
data and write this data out, going back to sleep until it is needed.
For capture the thread sleeps until the capture size has been met,
then reads out that many samples and sends a message telling the
client that there is data to be read in shared memory.

Playback Flow:
Client -> SHM -> alsa_iodev -> ALSA buffer

Capture Flow:
ALSA buffer -> alsa_iodev -> shm -> Client

BUG=chrome-os-partner:6382
TEST=alsa_io_unittest plus build/run on mario, lumpy64, and host amd64

Change-Id: Iebf3cda566986a142c86045364cb3c9d20fd1ac4
Signed-off-by: Dylan Reid <dgreid@chromium.org>