CRAS: Add system settings module.
This module is initialy used to hold the system volume.  This volume
is applied by the active output device, and is independent of stream
volumes that are managed in cras_rstream.  The system volume sets the
overall volume level for the system this is applied to all plaing
streams post-mix.

BUG=chromium-os:25411
TEST=new cras_system_settings_unittest.

Change-Id: I5d8a6673ff8314af141949d9fcb57ad6a27fef99
Signed-off-by: Dylan Reid <dgreid@chromium.org>