CRAS: Store socket files somewhere reachable by all.
This will allow the server and client to be run by different users
(which is how it will work for security purposes on ChromeOS).  The
common place to store the files will be "/var/run/cras/".  This can
easily be turned into a run-time configurable option, but there is
currently no need.

Set permission more restrictively on the socket files, limit access
to group members.

BUG=chromium-os:25409
TEST=build and run on Mario.

Change-Id: I08e42388e0eedb65bee6ba7ecbef58513d3c3fdf
Signed-off-by: Dylan Reid <dgreid@chromium.org>