CRAS: Set all volume elements.
Don't stop setting the volume elements when we've applied enough gain.
If volume gets set in a jump say moving from 10 to 90, secondary
controls would have been left in the state for 10.  Now we will set
the volume for each control whenever the volume is changed to make
sure we get the correct amount of attenuation on the output.

BUG=none.
TEST=manual, boot and run cras, watch alsa controls change as the
volume is adjusted.

Change-Id: I67d16633ff5e088a742c4e3c01f30a4698f471ec
Signed-off-by: Dylan Reid <dgreid@chromium.org>