CRAS: Add stream from client thread.
When adding a stream, the client needs to wait for the server to pass
back the client id.  We can only wait for a message from the client
thread so move the stream creation there and add a check and wait
before sending the add stream message.

Chrome creates two clients rapid fire when playing a video which
exposed this bug.

BUG=none
TEST=build and run on ChromeOS, load youtube and hit refresh often to
test client creation.

Change-Id: I378650641c5cddbf3f93ed2cea1cb1027b4f085b
Signed-off-by: Dylan Reid <dgreid@chromium.org>