CRAS: Make alsa_io keep a list of outputs.
Create a list of output nodes per iodev.  Allow for one of those nodes
to be set as the active node.  Only one output can be made active at a
time.  Examples of outputs are headphones, speakers, and IEC958.
These can be any switch that allows for a unique output path.

Change-Id: Ia5326a70dd6faf042059c7728fd91c77003bc7f0
Signed-off-by: Dylan Reid <dgreid@chromium.org>