CRAS: Add empty iodev.
Add an empty iodev.  This iodev will be the fall back for when no
devices are attached to the system.  All ChromeOS devices have built
in audio so this situation is only possible at first boot before the
drivers are loaded.  For ChromiumOS installs it could be the norm
until an external sound card is plugged in.

BUG=none
TEST=none

Change-Id: I8fdbce053f098e2063f84eb52cef70f429828a74
Signed-off-by: Dylan Reid <dgreid@chromium.org>