CRAS: Auto route audio to new devices.
Change the default device when a new device is added.  Add a param to
add_[out|in]put so that this behavior can be disabled.  For alsa cards
only route to the first device (device 0).  All devices on a card are
inserted at the same time so we have to pick one, and the first is as
good as any for now.  JACK detection will be added next and that will
allow for already present devices to become the default and for smart
selection of which device to route to when a new card is inserted.

Add auto-route devices to the beginning of the device list, this will
cause them to be used in the order they were inserted.  When the
current device is remove, fall back to the beginning of the list.

Example:

Insert device A
Insert device B
Add stream -> plays to device B.
Remove device B -> stream moves to device A
Insert device C -> stream moves to C.
Remove device A
Insert device D -> stream moves to device D
Insert device E -> stream moves to device E
Remove device D
Remove device E -> Stream moves to device C
Remove device C -> Stream goes to empty device.
Insert device A -> Stream moves to device A.

BUG=chromium-os:533
TEST=updated unit tests + plug and unplug USB audio devices and see
audio routed to them.

Signed-off-by: Dylan Reid <dgreid@chromium.org>
Change-Id: I1f316b67a0192d183dfe60b6b7268c8b6d42244f