CRAS: Add an alsa plugin.
This alsa plugin will allow applications to only talk to ALSA to play
nicely with CRAS apps.  The current GTalk plugin talks to ALSA through
its own WebRTC stack so this is needed to make that plugin play nice
with the rest of Chrome.

BUG=chromium-os:25409, chromium-os:25410
TEST=Build plugin configure it as default ALSA device and run GTalk.
Notice that volume can be adjusted through CRAS.

Change-Id: I9ccc127afe400915243d0c28efff79fafa0545f6
Signed-off-by: Dylan Reid <dgreid@chromium.org>