CRAS: system, allow multiple volume callbacks.
Change the volume and mute callbacks so that more than one user can be
registered for a callback at a time.  Previously only the active io
device asked to be notified of the system volume, however now there is
a need to be able to pass that volume back to clients.

BUG=chromium-os:25409
TEST=updated unit tests.

Change-Id: Iebaf896efbcff87136249822b4b4d96fcd9e9866
Signed-off-by: Dylan Reid <dgreid@chromium.org>