CRAS: Track capture gain and mute in system state.
Pass in a gain value in dB * 100.  Most often this will be a positive
value applying a gain to the input.  Use a dB value here instead of an
index to provide fine grained control needed by AGC algorithms used in
GTalk.

Subsequent changes will add a message to adjust this and plumb it up
to the ALSA capture controls.

BUG=chromium-os:25410
TEST=updated unit tests.

Change-Id: I6078d9daad9b4e58c20fd7f4772ec0b3c51fe0bd
Signed-off-by: Dylan Reid <dgreid@chromium.org>