CRAS: mixer - Add ability to set capture gain.
Set the capture gain through the alsa mixer.  This allows the input
gain level to be adjusted.  Look for elements of type, "Mic Boost",
"Capture", and "Digital Capture".  The boost control will be used
first to apply as much needed gain as we can in the analog path, then
capture and finally digital capture.  Any controls that aren't present
in the mixer will be skipped.

BUG=chromium-os:25410
TEST=updated mixer UT.

Change-Id: I5a1ec24415ee3c79963ef21ad2c4af9954d372d7
Signed-off-by: Dylan Reid <dgreid@chromium.org>