CRAS: client - add retries when connecting to the server.
Add an api "cras_client_connect_wait" that tries to connect to the server 4
times before quitting.  This will fix the issue where a client is started
momentarily before CRAS is.  Also move factor out the connect logic so it can be
reused by a future change.  Ignore SIGPIPE as the pipe errors will be handled
when read returns a failure.

BUG=chrome-os-partner:9383
TEST=build and run test client.

Change-Id: I13808f5317f6efd9b4de3812190d9988c78a9354
Signed-off-by: Dylan Reid <dgreid@chromium.org>