CRAS: client - bounds check num_frames in playback cb.
Format conversion calculates the number of source format frames to
request based on the number of hardware frames needed. For example, when
calculating the amount of 48k frames needed to mix with a given number
of 44.1k frames, the requested number could end up larger than the
buffer size by one due to rounding. To avoid this check that the number
of frames requested is never bigger than the buffer size.

Change-Id: Iea0a707de93d2b0b43fb9c161ae3a15c54b8f378
Signed-off-by: Dylan Reid <dgreid@chromium.org>