CRAS: client - Add function to get bytes in a frame for a format.
Add a function that can get the number of frames given an audio format.
This replaces the cras_client_get_bytes_per_frame with
cras_client_format_get_bytes_per_frame.  Using the new function instead
of the old removes access to the streams linked list from the audio
thread and avoids a search for the stream.  Deprecate the old function,
it will be removed as soon as chrome is switched to use the new
interface.

BUG=chromium:125750
TEST=cras_test_client still works.

Change-Id: Ib416c648cc26e3f7c48da58ef703e55a2ae2db6c
Signed-off-by: Dylan Reid <dgreid@chromium.org>