CRAS: Add priority to iodevs.
Give each device a priority.  This priority will be used to decide which device
to make the default at startup.  For the initial decision there is no ordering
information to indicate where the user wants audio.  We will default to external
devices, assuming they have been attached for a reason.  A follow up CL will use
this to sort the devices at startup.

BUG=chromium-os:533
TEST=updated unit tests.

Change-Id: I9012f10d8de29c18c955b0bdce22948c5735359f
Signed-off-by: Dylan Reid <dgreid@chromium.org>