CRAS: rclient - Sort iodevs before adding first stream.
Before the first stream is added, sort the iodevs.  This achieves a
reliable initial priority of devices.  After boot, there is no way to
know which devices were attached when.  Instead we give priority to
external devices.  After the first stream is added, any new devices
attached will go to the head of the list and remain there until they are
unplugged or another device is attached.

BUG=chromium-os:533
TEST=updated rclient_unittest

Change-Id: Ie46ac02920e5b9864dd8af2e7433a379ff95f0e8
Signed-off-by: Dylan Reid <dgreid@chromium.org>