CRAS: shm - Fix cras_shm_get_avail_curr_buffer.
This function was unused until recently.  It was checking the wrong
buffer (using read_idx instead of write_idx).  Change the function to
enforce that a buffer is only written to once before it is read.  The
server only requests data when there is a buffer available so if
write_offset is non-zero in a client callback there has been an error
(alsa_io, checks on the server side before requesting data).  In this
case return zero so nothing is written.

BUG=none
TEST=updated shm unit test, build and run on ZGB, youtube, gtalk and
WebAudio drum kit demo.

Change-Id: Idd1aade02b2c23c76fc5e79ff57c3ba9bf6be7c8
Signed-off-by: Dylan Reid <dgreid@chromium.org>