CRAS: client - Give fmt_conv_create the real max_frames.
When up-sampling (44.1 to 48) the number of frames the server wants will
be more than the number the client requests.  Configure the format
converter to handle the greater of the two frame counts.  The server
specified the format-converted frame count in the shm region.  Take the
greater of the number in shm and the local number stored in the stream
config.

BUG=chromium:130093
TEST=run 48k audio (drum kit demo), and 44.1 (pic a wav file).

Change-Id: Ife614b6e0a6b3f18ff0e1e31fd42f8ec1f7a1d54
Signed-off-by: Dylan Reid <dgreid@chromium.org>