Add alsa-$(board).conf files for select boards.
Some boards need their model name passed to the hda driver.  That was
done in alsa-utils before we switched back to upstream.  This file is
optional, if a board needs to pass options to the alsa driver it will
create one.  The file is copied to /etc/modprobe.d on the target.

BUG=chromium-os:31063
TEST=build stumpy and alex, boot and check that the correct verbs have
been loaded by checking "amixer -c0 controls" and
"/proc/asound/card0/codec#0".

Change-Id: I67725fe9532c95b48fe1d23968cb0024d941e3e8
Signed-off-by: Dylan Reid <dgreid@chromium.org>