CRAS: alsa capture - Fix sleep times if target is undershot.
Set the next timeout correctly if the audio thread wakes up before
samples are ready.  Before, this would set an incorrect timeout,
resulting in waiting for longer than needed and causing the next two
wakes to be close to each other.  The results of this are that wakes
during low latency record are much more regular.

BUG=chromium-os:9539
TEST=Record on lumpy, use pytimechart to visualize wakeups from cras.

Change-Id: Iee821bb35c53d71e63f3323bfaaa4e5d97b15668
Signed-off-by: Dylan Reid <dgreid@chromium.org>