CRAS: alsa capture - Sleep a few extra frames.
Sleep for more frames than the minimum needed during capture.  This
reduces the possibility of waking up too early and having to perform a
short sleep before waking up again.  Introduce a correction factor that
is adjusted when the sleep is over or under shot.  This correction
factor will track to a value that causes wakes to happen when the
desired amount of data (plus the extra frames added here) is available.

BUG=none.
TEST=update unittest, run pytimechart on lumpy while recording audio,
check that the wake interval is stable.

Change-Id: Id6d33505ba6f25f3903fa20d87f6defab8bdaff8
Signed-off-by: Dylan Reid <dgreid@chromium.org>