ADHD: Remove unused code
Now that cras handles the state of 'active' hardware, there is no
reason for gavd to continue maintaining that information.  Remove
almost all code related to broadcasting & detecting hardware events.

The only hardware code that is left now is the automatic switching of
output to the headphone on Arm machines.

BUG=none
TEST=Code removal;  nothing else impacted.
Signed-off-by: Taylor Hutt <thutt@chromium.org>

Change-Id: Ia9bbe8fed9b171974ccbbf6ef8008423bc9b0343