CRAS: alsa mixer - use max device volume as offset in dB.
Use the max volume reported by the device as an offset for the volume
adjustments.  Add up the volumes of all the main controls and use that
number as a starting point.  Do this so that USB devices whose volumes
commonly range from zero to a positive dB value can be adjusted.  This
will have no affect on any chromebooks internal volume curve as they all
have a max of 0.

BUG=chromium-os:31755
TEST=Update UT and attach USB speakers and adjust volume.

Change-Id: I9175d3b53281ed9470b41f9dfe5d0552f2214a74
Signed-off-by: Dylan Reid <dgreid@chromium.org>