CRAS: iodev_list - don't route away from plugged device.
When sorting the device list, don't move streams that are not on the
default device.  Streams move to non-default devices when a jack is
plugged and when a device has a plugged jack it should remain the
default until it is unplugged.  Fixes an issue with routing away from
HDMI/DP devices at startup.

BUG=chromium-os:20619
TEST=updated UT, boot stumpy with DP plugged, head audio from DP device.

Change-Id: I245c9351cb140f4d0c6baee68fd13dd16d757777
Signed-off-by: Dylan Reid <dgreid@chromium.org>