CRAS: Ignore cb_threshold from clients.
Allowing clients to set this value doesn't provide any flexibility and
chrome always sets it to the same thing.  Switching to use the
min_cb_level instead as it will provide the same latency, but give more
time for the callback to be handled which makes USB's lower buffer level
play nicely.

BUG=chromium:133843
TEST=webrtc loopback test page with sound blaster arena headset.

Change-Id: Id118c8e346b5febc5904acf76b145ed01df04108
Signed-off-by: Dylan Reid <dgreid@chromium.org>