Add volume curves for stout and butterfly.
Volume control did not work correctly on stout and butterfly due to dB
scale mismatch. These volume curves are based on the stumpy volume
curve, but scaled to stout + butterfly dB scale.

BUG=chrome-os-partner:10515
TEST=Adjust volume slider on butterfly + stout platforms, verified that
audio does not become muted early and that volume scale is reasonable on
both speaker and headphone jack.

Change-Id: Id6c7d27aac5afaec141933c7e9ad24ef61d93fc7