Only do hardware setup when control device is available
Only do the device add and alsactl setup on startup when the control
device has been added. The way udev handles this is by posting both an
add and a change event when the device is being added, so we only want
to do this on the change event.

However, to make things complicated, if cras is started after the device is
available, i.e. if cras is restarted, then we will only see the add events.
The way this is handled is that udev will set a SOUND_INITALIZED=1 attribute
on the main card node for those cases, so we check for that in the add path
and if that's seen, then we do the device addition and hardware reset there.

This has been tested on daisy, and seems to remove the race where we sometimes
try to do the hardware reset before the control node is available.

Note that this changes system setup on all platforms a bit; ALSA devices will no
longer be registered until the corresponding control device has been registered
(and reported stable through the above method).

BUG=chrome-os-partner:10957
TEST=boot system, make sure audio works and make sure /var/log/messages has an entry for:
set_factory_default: init card '0' to factory default from the last boot.

Change-Id: I798a0f438d94a415979da8e92ef7c988bb510bee