CRAS: Add USB output device blacklist.
Add the ability to black list certain USB output devices.  A device is
specified by a vendor/product ID pair and the device index in the card.
No support for USB inputs yet as there is no known need.  The need to
blacklist output devices is explained in the bug.

BUG=chromium-os:32278
TEST=new device_blacklist_unittest

Change-Id: I489c0b0047ca580855860f193635fbb6987f8230
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27607
Reviewed-by: Olof Johansson <olofj@chromium.org>