factory-default: Set Mic Capture volume for stumpy/lumpy.
Default to setting the Mic Capture gains to +20dB for stumpy and lumpy.
The main capture volume is adjusted by the AGC and can be used to
attenuate this, but it can't be used to gain more than 12dB.  Setting
these higher moves the default capture gain to a more useable range.

BUG=chromium-os:30670
TEST=desktopui_AudioVideoFeedback

Change-Id: Ie6f03852e63ab385802b6e0bc38f64ad1d0650d3
Signed-off-by: Dylan Reid <dgreid@chromium.org>