CRAS: alsa_card - route to new default devices.
Use the new priority scheme from iodev list to choose the correct device to
route to.  This removes the assumption that a jack plugged event should always
route to the associated device immediately.  When HDMI is plugged in we can now
avoid routing audio away from headphones.

BUG=chromium-os:32730
TEST=updated alsa_io unittest.  plug in headphones then hdmi audio continues on
headphones, goes to HDMI when headphones are detached.

Change-Id: I36c7ff79559d19db696b33d54c499f85365c92b0
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27947