CRAS: Set plug time for USB devices when attached.
When a USB device is attached treat that as if a headphone had been
plugged in.  The USB headphones being plugged in should behave in the
same way as plugging in analog headphones.  This causes them to sit at
the appropriate "most recently plugged" priority.

BUG=chromium-os:32650
TEST=updated alsa mixer unittest, plug and unplug USB headphones and
HDMI, make sure audio returns to correct place after detach.

Change-Id: I0bbabdf858ab05260a3e90931855d69de26b462c
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27859