CRAS: shm - Make get_write_buffer check how much can be written.
Change cras_shm_curr_write_buffer to cras_shm_get_writeable_frames.  The
new function still returns a pointer to the location to write, but also
fills in a parameter with the count of frames that can be written.  This
mirrors the behavior of cras_shm_get_readable_frames.  Checking both of
these in one place allows bounds checking of writes to the buffer.

BUG=chromium-os:33123
TEST=manual, youtube playback, youtube.com/my_webcam.

Change-Id: I75be87a961f261f6a53e725fd217a9a82edd2248
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29294