CRAS: alsa-mixer - Save max volume for each output.
The max volume was being saved for card-wide controls such as Master or
PCM, but not for individual outputs like Headphone or Speaker.  Daisy's
outputs each go above zero dB, there is now a need to track each
output's max volume.  Save the max volume setting when discovering the
output and add it in when applying it.

BUG=chrome-os-partner:12109
TEST=updated unit test, play youtube on daisy and adjust volume.

Change-Id: I595211f88fa4a4103b28127884550bc19cb70399
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29305