CRAS: system state - factor out server state struct.
Move the members of the system state struct that will later be shared
with clients to their own sub-struct.  A later change will move this
server_state struct to a shared memory region.  Version the structure in
case it needs to be updated and clients are forced to handle multiple
versions.

BUG=chromium-os:32347
TEST=unit tests still pass.

Change-Id: I0584b78cb355902ec6f40399facc4562313d76b6
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29858
Reviewed-by: Olof Johansson <olofj@chromium.org>