CRAS: iodev_list - Store iodev info in system state shm region.
In addition to sending the iodev list to clients, also update the shm
region to contain the same info.  A subsequent change will change the
clients to get the information from here and remove the message.  Ignore
the result fo updating the shm area, a failure here won't hurt the
server at all.

BUG=chromium-os:32347
TEST=updated unittest, 'cras_test_client --dump_server_info'

Change-Id: I9bdd6c9e2850ebbbece32518e58b99467c27056f
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29861
Reviewed-by: Olof Johansson <olofj@chromium.org>