CRAS: client - Use server_state shm instead of messages.
This removes a lot of extra messaging from server to client and between
the client thread and the client.  Use the shared memory region to get
server data instead of querying the running client thread for it.

The change to the test client is necessary since the 'get_system_volume'
call no longer blocks waiting for the server.  Also ensure that the shm
region has been updated from the alsa control.

BUG=chromium-os:32347
TEST=cras_test_client to get server info and play audio.  Check
chrome://system cras tab and make sure it is still correct.

Change-Id: I1699e5423f0c360982dac901714e8fdb3f70ae20
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29864
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>