CRAS: client - Add helper function to check if device is plugged.
This helper will check the output device based on the given name and
return true if the device is plugged.  For internal devices this means
that the jack associated with the device is present (HDMI/DP or
Headphone).  For USB devices if the device is attached, plugged is true.

BUG=chrome-os-partner:10413
TEST='cras_test_client --check_output_plugged "HDA Intel PCH: Cirrus
Analog"' on Lumpy.

Change-Id: I8e0cc9a3ab7664ae6176bc1896f42758817b265b
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30383