CRAS: alsa io - Close alsa on error.
When an error is reported from alsa, close the device.  There have been
instances on Daisy where USB audio devices disappear, but it takes a
while for the udev notification to happen.  Without this change CRAS
would end up in a loop getting errors from the device until notified by
udev that it is gone.  Instead close it immediately but wait for the
notification to clean up the rest of the card.

BUG=chrome-os-partner:12785
TEST=alsa io unit test, plug and unplug USB audio on snow, check log for
long strings of error messages (should no longer be there, now just one
per removal).

Change-Id: Ia8dcb850d8b5f4bf6f18cbb6a1104492094b19db
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30818