CRAS: iodev list - Handle system volume changes.
Register callbacks for system volume and capture gain changes.
Callbacks for these are currently registered by the individual iodevs,
but then need to be cleaned up when the devices are removed or go idle.
The callbacks only matter for the active device, and the iodev_list
knows which device that is.  Use this information to call back the
correct device.

BUg=chrome-os-partner:12788
TEST=iodev_list_unittest

Change-Id: If18d3ee4135286f65066f0b13cd91a2ef7f79201
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30911