CRAS: gpio jack - Input jacks can't detect "Headphone Jack".
Change the regex used to detect input jacks.  Matching Headphone jacks
on the input path leads to input devices acting on jack changes meant
for the output.  If a device has a shared jack, it will have to rename
it so it can be handled properly.  This change may break input switching
on Waluigi, but input switching didn't work on that board to begin with.

BUG=chrome-os-partner:13081
TEST=plug/unplug mic on Daisy.

Change-Id: Ic7ff265ecb5dfff0873972cd5b8b4030ccb019a3
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31297