CRAS: alsa io - Don't delete stream on open error.
If there is an error opening the alsa device, don't blame the stream an
delete it.  Allow the stream to remain and be moved to another device
when this one disappears or the other device becomes available.

Initialize the handle pointer as it is now possible it won't be set in
open.

BUG=chromium-os:33624
TEST=attach headphones and USB speakers, play audio, then suspend,
unplug both while suspended and resume.  Audio should route back to
built-in speakers.

Change-Id: Ieef62bc2aa125992190cb53f995b27be3f9898c3
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31317