CRAS: alsa plugin - output the actual delay from alsa plugin to CRAS.
This change adds the delay in CRAS to what we're now reporing. Right now
alsa pcm plugin maintains a copy of the latencies to/from CRAS, which is
derived by the timestamp passed into client callback and the buffer
count transfering between CRAS and alsa plugin.

BUG=chromium-os:33366
TEST=Run program to measure the delay through CRAS and make sure it's
the same as we reported.

Change-Id: I8595b12ce5730fbacf2838898433c474dff1a877
Reviewed-on: https://gerrit.chromium.org/gerrit/31456
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>