CRAS: sys state - Track the number of active streams.
Keep a count of the number of audio streams attached to the server.
This requires adding a lock, as streams can be removed from the audio
threads if they report an error.  Also track the time the last thread is
removed from the system.  This timestamp will be used by the power
manager to measure how long audio has been idle.

BUG=chromium-os:32500
TEST=updated system_state_unittest

Change-Id: Ifa9e1294713eebeac28e6cdffabe9cd770edb5c9
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/32449