link asount.state and ucm config - toggle DSP on headphone/mic (un)plug.
Add a UCM config for link. Toggle DSP controls when headphone and mic
are plugged/unplugged.

We don't want to use the DSP effects for headphone output, and the input
blocks we want to run are different when using an external mic.

Change the asound.state to use the configuration we want when
headphone/mic is unplugged.

BUG=chrome-os-partner:14154
TEST=use alsamixer -c 0 to verify the setting are changed when headphone is
plugged/unplugged.

Change-Id: I9ac27208267e2fb0993e2c10780e8cd12508ce10
Reviewed-on: https://gerrit.chromium.org/gerrit/33775
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Ready: Chih-Chung Chang <chihchung@google.com>
Tested-by: Chih-Chung Chang <chihchung@google.com>