CRAS: iodev - update supported frame rates.
Before setting the format of the iodev, check which formats are
currently supported.  On Daisy, the input and output must be run at the
same rate if used concurrently.  Updating the supported rates before
opening the device makes sure that a rate/channel count that will work
is chosen.

BUG=chrome-os-partner:15126
TEST=play youtube (44.1) then start a hangout(48k) check that the mic
works in the hangout.

Change-Id: I3430dc0f8fb13495dd6e385732bc8f46a4fc9619
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/35369
Reviewed-by: Hsinyu Chao <hychao@chromium.org>