CRAS: alsa_io - Add get/put buffer helpers.
Pull out getting and returning mmap buffers from alsa to helper
functions.  This will aid in refactoring the logic in the read/write
functions to be alsa-independent.  Once it is alsa independent it will
be able to be reused across other types of io devices.
Rework some types so that no alsa-specific types are used.

BUG=chromium-os:35272
TEST=alsa_io_unittest

Change-Id: I81306a8aa54c599885ce6cbb5e8c1e6f7b673ed5
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39834
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>