CRAS: bluetooth - track default adapter
Communicate with the BlueZ bluetooth daemon via D-Bus to obtain the
object path of the system default bluetooth adapter. Use signals from
BlueZ to track changes in the default as adapters are removed and
inserted. Also track signals from the D-Bus daemon itself as the
bluetooth daemon is stopped and started.

BUG=chromium-os:21369
TEST=src/bluetooth_unittest

Change-Id: I0a90335a9062da77fe008f76e41916821ae7761a
Signed-off-by: Scott James Remnant <keybuk@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/35710
Reviewed-by: Dylan Reid <dgreid@chromium.org>