CRAS: iodev - Create thread outside of iodev.
Instead of having the iodev create its own thread, create the thread and
pass it to the iodev.  This will allow for iodevs to share the same
thread, enabling pairing input and output thread for unified IO.

This moves a bunch of code from alsa_io/iodev into audio thread as the
functionality will be common to all iodevs, not just alsa.

BUG=chromium-os:35272
TEST=updated unittest, plus, GTalk and YouTube on device.
Signed-off-by: Dylan Reid <dgreid@chromium.org>

Change-Id: I90bcd53243c8564922d0e716ed5f3e176aff566e
Reviewed-on: https://gerrit.chromium.org/gerrit/40275
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Dylan Reid <dgreid@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>