CRAS: DSP - Add expression parsing and evaluation functions and unit tests.
The expression facility will be used to determine whether a plugin needs
to be enabled. For example, if we have this line in the ini file for
the DRC plugin:

disable=(or disable_drc (equal? output_device "HDMI"))

then the plugin will be disabled when this expression is evaluated to
true (that is, when the variable "disable_drc" is true or the variable
"output_device" equals to "HDMI".

The expression syntax generally follows the syntax in the Scheme
programming language, but we only use a (very small) subset of it.

BUG=chromium-os:36300
TEST=make check
Change-Id: If57f4f0ba2863ef5ed2d7592d3563a6dd5bee05d
Reviewed-on: https://gerrit.chromium.org/gerrit/38379
Commit-Ready: Chih-Chung Chang <chihchung@chromium.org>
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>
Tested-by: Chih-Chung Chang <chihchung@chromium.org>