CRAS: Add dumper utilities and unit tests.
A dumper is used to obtain human-readable text from a component for
debugging and informational purposes. The caller can pass different kinds
of dumpers to a component, and the text will go to different places.

Currently there are two dumpers, one goes to syslog and another goes to
a memory buffer.

BUG=chromium-os:36300
TEST=make check
Change-Id: Iee669e06b013517ca1c7f8ac69af4b5c081a051e
Reviewed-on: https://gerrit.chromium.org/gerrit/38378
Commit-Ready: Chih-Chung Chang <chihchung@chromium.org>
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>
Tested-by: Chih-Chung Chang <chihchung@chromium.org>