CRAS: client - Add API to create unified io stream.
Add a callback type and an API for unified audio.  This API will be
used to create a stream that processes input and output audio in one
callback.  The callback for the client is similar to the existing
audio callback, but specifies two buffers, one input and one output
for audio.  When the callback is run captured samples can be read from
the buffer, processed and fed back to the playback buffer to minimize
latency.

BUG=chromium-os:35272
TEST=updated test client.

Change-Id: I9a76b5637537c729057ddbf55cfab8824835de2a
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/40966