CRAS: Allow input and output device for each thread.
For unified audio there is a need to have one thread service both an
input and an output device synchronously.  In this case the thread
will be initialized with the input device, then the output device will
be added.  This doesn't yet handle moving streams already attached to
the output device.

BUG=chromium-os:35272
TEST=unit tests pass, and test unified audio with test_client.

Change-Id: Id98e248e9c0643dd37e00fb652c6bbf144546f5f
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/40968