CRAS: rclient - Set output dev format better.
The output device format shouldn't be set until after the thread is
created for unified streams.  If the output device already contains
streams, its format will be reset when the streams are removed, set
the format of the iodev later in this case. It will be set in the same
place for output-only streams.

BUG=chromium-os:35272
TEST=play output then add unified stream, hear output continue.

Signed-off-by: Dylan Reid <dgreid@chromium.org>
Change-Id: Iff30a4818831066e367dbf6d97e9948307378ca8
Reviewed-on: https://gerrit.chromium.org/gerrit/41134